# [SCFD-3041] Improve validation and translation performance by removing duplicate entity warning