--- a/tools/performance_profile/profiler.py
+++ b/tools/performance_profile/profiler.py
@@ -1,5 +1,6 @@
 import cProfile
-import time, json
+import json
+import time
 
 from flow360.component.simulation.services import (
     simulation_to_case_json,
@@ -11,7 +12,7 @@ from flow360.component.simulation.services import (
 start_time_glb = time.time()
 start_time = time.time()
 with open("./data/large_simulation.json", "r") as f:
-    params_as_dict= json.load(f)
+    params_as_dict = json.load(f)
 end_time = time.time()
 print(f"Execution time: {end_time - start_time} seconds [json.load()]")
 start_time = time.time()
@@ -27,6 +28,7 @@ print(f"Execution time: {end_time - start_time} seconds [translation Case+SM+VM]
 
 def translation_wrapper():
     simulation_to_surface_meshing_json(params, {"value": 100.0, "units": "cm"})
+
 
 def validation_wrapper():
     _, _, _ = validate_model(params_as_dict, "SI", "Geometry")