… tools/performance_profile/data/large_simulation.json
… tools/performance_profile/profile_translator.prof
… tools/performance_profile/profile_validation.prof