[DEV001]
[BEFORE]
```
Execution time: 0.013393163681030273 seconds [json.load()]
[15:29:14] INFO: using: SI unit system                                                                                                      
Execution time: 1.751427412033081 seconds [validate_model]
[15:29:15] INFO: using: SI unit system                                                                                                      
           INFO: using: SI unit system                                                                                                      
           INFO: using: SI unit system                                                                                                      
Execution time: 6.627842664718628 seconds [translation Case+SM+VM]
[15:29:20] INFO: using: SI unit system                                                                                                      
           INFO: using: SI unit system 
``` 
[AFTER]                                                                      
```
Execution time: 0.012851238250732422 seconds [json.load()]
[15:28:48] INFO: using: SI unit system                                                                                                      
Execution time: 0.21953296661376953 seconds [validate_model]
           INFO: using: SI unit system                                                                                                      
           INFO: using: SI unit system                                                                                                      
           INFO: using: SI unit system                                                                                                      
Execution time: 2.1067793369293213 seconds [translation Case+SM+VM]
[15:28:50] INFO: using: SI unit system                                                                                                      
[15:28:51] INFO: using: SI unit system 
```